backend/requirements.txt
# Backend dependencies with versions matching current environment
# Updated based on pip list output from current environment

flask==3.0.0
flask-cors==6.0.1
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.7.1
werkzeug==3.1.1
pycryptodome==3.23.0
pymysql==1.1.2
cryptography==43.0.0
flask-mail==0.10.0
yfinance==0.2.63
defeatbeta-api==0.0.33
pandas==2.2.3
numpy>=1.26.4
google-generativeai==0.8.5
python-dotenv==1.2.1
python-dateutil==2.9.0
requests==2.32.3
akshare==1.17.26
alpha_vantage==3.0.0
tushare==1.4.19
apscheduler==3.11.2
stripe==14.1.0
supabase==2.25.1
psycopg2-binary==2.9.11
pydantic==2.12.4
scipy==1.16.2
tigeropen==3.4.3
gunicorn==23.0.0